Fixed code review comments raised from RWA-807
