+++ src/test/java/uk/gov/hmcts/reform/wataskmanagementapi/services/CFTTaskDatabaseServiceTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.reform.wataskmanagementapi.cft.entities.TaskResource;
 import uk.gov.hmcts.reform.wataskmanagementapi.cft.repository.TaskResourceRepository;
@@ -14,7 +15,11 @@ import java.util.UUID;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.spy;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
@@ -69,5 +74,26 @@ class CFTTaskDatabaseServiceTest {
         final TaskResource actualTaskResource = cftTaskDatabaseService.saveTask(someTaskResource);
 
         assertNotNull(actualTaskResource);
+    }
+
+    @Test
+    void should_return_task_resource_by_specification() {
+        TaskResource taskResource = spy(TaskResource.class);
+        Mockito.when(taskResourceRepository.findOne(any())).thenReturn(Optional.of(taskResource));
+
+        final Optional<TaskResource> taskBySpecification = cftTaskDatabaseService.findTaskBySpecification(any());
+
+        assertNotNull(taskBySpecification);
+        verify(taskResourceRepository, times(1)).findOne(any());
+    }
+
+    @Test
+    void should_return_empty_task_resource_by_specification() {
+        Mockito.when(taskResourceRepository.findOne(any())).thenReturn(Optional.empty());
+
+        final Optional<TaskResource> taskBySpecification = cftTaskDatabaseService.findTaskBySpecification(any());
+
+        assertTrue(taskBySpecification.isEmpty());
+        verify(taskResourceRepository, times(1)).findOne(any());
     }
 }
